Release 1.0.6
🩹 Enhance logout behavior

🐛 Fix start session for H-Case

✅ Add Unit test - coverage 40%

🚨 Fix Unit tests lints

🚀 Bump SDK version

+++ Sources/Userpilot/Experiences/Theme/Models/ThemeContent.swift
@@ -153,7 +153,7 @@ internal struct ExperienceTheme: Decodable {
     }
 
     var backdropOpacity: CGFloat {
-        CGFloat(CGFloat(backdrop?.opacity ?? ThemeHandler.DefaultValues.dimSlideOutDegree) / 100)
+        CGFloat(backdrop?.opacity ?? ThemeHandler.DefaultValues.dimSlideOutDegree) / 100
     }
 
     var backdropBackground: UIColor {
@@ -320,7 +320,7 @@ internal struct SurveyTheme: Decodable {
     }
 
     var backdropOpacity: CGFloat {
-        CGFloat(CGFloat(backdrop?.opacity ?? ThemeHandler.DefaultValues.dimSlideOutDegree) / 100)
+        CGFloat(backdrop?.opacity ?? ThemeHandler.DefaultValues.dimSlideOutDegree) / 100
     }
 
     var backdropBackground: UIColor {
